feat: 칭찬 랭킹 조회, 칭찬 삭제 구현

+++ src/main/java/com/ssginc/secretgarden/domain/compliment/dto/response/ComplimentRankingResponse.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ComplimentRankingResponse {
-    private Integer rank;
+    private Long rank;
     private Integer memberId;
     private String name;
     private Integer companyId;
